--- 05_student_test_shell/src/main/java/ru/otus/spring/events/EventsPublisher.java
@@ -1,5 +1,0 @@
-package ru.otus.spring.events;
-
-public interface EventsPublisher {
-    void publish();
-}
