--- 05_student_test_shell/src/main/java/ru/otus/spring/events/HalfAGlassOfWaterEvent.java
@@ -1,15 +1,0 @@
-package ru.otus.spring.events;
-
-import lombok.Getter;
-import org.springframework.context.ApplicationEvent;
-
-public class HalfAGlassOfWaterEvent extends ApplicationEvent {
-
-    @Getter
-    private final String payload;
-
-    public HalfAGlassOfWaterEvent(Object source) {
-        super(source);
-        payload = "Осталось половина стакана воды!!!";
-    }
-}
