--- 05_student_test_shell/src/main/java/ru/otus/spring/events/HalfAGlassOfWaterEventPublisher.java
@@ -1,17 +1,0 @@
-package ru.otus.spring.events;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class HalfAGlassOfWaterEventPublisher implements EventsPublisher {
-
-    private final ApplicationEventPublisher publisher;
-
-    @Override
-    public void publish() {
-        publisher.publishEvent(new HalfAGlassOfWaterEvent(this));
-    }
-}
