+++ 05_student_test_shell/src/main/java/ru/otus/spring/service/interfaces/QuestionsService.java
@@ -3,8 +3,7 @@ package ru.otus.spring.service.interfaces;
 import java.io.IOException;
 
 public interface QuestionsService {
+
     void takeExam();
-    String getStudentName();
-    void setStudentName(String studentName);
 
 }
