+++ 05_student_test_shell/src/main/java/ru/otus/spring/service/interfaces/StudentLoginService.java
@@ -1,0 +1,8 @@
+package ru.otus.spring.service.interfaces;
+
+public interface StudentLoginService {
+
+    String getStudentName();
+    void setStudentName(String studentName);
+
+}
