+++ 05_student_test_shell/src/main/java/ru/otus/spring/shell/ApplicationEventsCommands.java
@@ -6,31 +6,34 @@ import org.springframework.shell.standard.ShellComponent;
 import org.springframework.shell.standard.ShellMethod;
 import org.springframework.shell.standard.ShellMethodAvailability;
 import org.springframework.shell.standard.ShellOption;
-import ru.otus.spring.events.EventsPublisher;
 import ru.otus.spring.service.interfaces.QuestionsService;
+import ru.otus.spring.service.interfaces.StudentLoginService;
 
 
 @ShellComponent
 @RequiredArgsConstructor
 public class ApplicationEventsCommands {
 
-    private final EventsPublisher eventsPublisher;
+
+    private final StudentLoginService stLoginService;
     private final QuestionsService questionsService;
+
 
     @ShellMethod(value = "Авторизация для студентов перед сдачей экзамена", key = {"l", "login"})
     public String login(@ShellOption(defaultValue = "Incognito") String studentName) {
-        questionsService.setStudentName(studentName);
+        stLoginService.setStudentName(studentName);
         return String.format("%s, добро пожаловать на сдачу теста. (test или help для начала прохождения)", studentName);
     }
 
     @ShellMethod(value = "Прохождение теста", key = {"t", "test"})
     @ShellMethodAvailability(value = "isPublishEventCommandAvailable")
-    public String publishEvent() {
-        eventsPublisher.publish();
-        return "Экзамен проведен. ('test' для повторного прохождения или 'exit' для выхода). Для студента: " + questionsService.getStudentName();
+    public String studentTest() {
+        questionsService.takeExam();
+        return "Экзамен проведен. ('test' для повторного прохождения или 'exit' для выхода). Для студента: " + stLoginService.getStudentName();
     }
 
     private Availability isPublishEventCommandAvailable() {
-        return questionsService.getStudentName() == null? Availability.unavailable("Сначала авторизуйтесь"): Availability.available();
+        return stLoginService.getStudentName() == null? Availability.unavailable("Сначала авторизуйтесь"): Availability.available();
     }
+
 }
