+++ 05_student_test_shell/src/main/resources/application.yml
@@ -11,6 +11,6 @@ spring:
 
 exam:
   correct-answers-to-pass: 3
-  locale-language: en
-  locale-country: US
+  locale-language: ru
+  locale-country: RU
   exam-file-name-csv: csv/exam-quastions
